+++ MyApplication/build.gradle
@@ -1,10 +1,10 @@
 buildscript {
     ext {
-        compose_ui_version = '1.2.0'
+        compose_ui_version = '1.4.3'
     }
 }// Top-level build file where you can add configuration options common to all sub-projects/modules.
 plugins {
     id 'com.android.application' version '8.0.1' apply false
     id 'com.android.library' version '8.0.1' apply false
-    id 'org.jetbrains.kotlin.android' version '1.7.0' apply false
+    id 'org.jetbrains.kotlin.android' version '1.8.21' apply false
 }
